backend/requirements.txt
# PICT Guard Backend Requirements
# Run: pip install -r requirements.txt

# Core Framework
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2
python-multipart==0.0.22
python-dotenv==1.2.1

# Database
motor==3.3.1
pymongo==4.5.0
dnspython==2.8.0

# Data Validation
pydantic==2.12.5
pydantic_core==2.41.5
annotated-types==0.7.0
email-validator==2.3.0

# Security & Auth
passlib==1.7.4
bcrypt==4.1.3
python-jose==3.5.0
PyJWT==2.11.0
cryptography==46.0.4

# QR Code Generation
qrcode==8.2
pillow==12.1.0

# Excel / File Processing
openpyxl==3.1.5
pandas==3.0.0
et_xmlfile==2.0.0

# HTTP & Networking
httpx==0.28.1
requests==2.32.5
aiohttp==3.13.3
urllib3==2.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11

# Email (SMTP)
# Uses built-in smtplib — no extra package needed

# Async Support
anyio==4.12.1
sniffio==1.3.1

# Utilities
python-dateutil==2.9.0.post0
six==1.17.0
typing_extensions==4.15.0
packaging==26.0
